+++ src/solvers/dgsem_tree/dg_3d_parabolic.jl
@@ -290,11 +290,13 @@ end
 
 function calc_parabolic_fluxes!(flux_parabolic,
                                 gradients, u_transformed,
+                                t,
                                 mesh::Union{TreeMesh{3}, P4estMesh{3}},
                                 equations_parabolic::AbstractEquationsParabolic,
                                 dg::DG, cache)
     gradients_x, gradients_y, gradients_z = gradients
     flux_parabolic_x, flux_parabolic_y, flux_parabolic_z = flux_parabolic # output arrays
+    @unpack node_coordinates = cache.elements
 
     @threaded for element in eachelement(dg, cache)
         for k in eachnode(dg), j in eachnode(dg), i in eachnode(dg)
@@ -307,17 +309,22 @@ function calc_parabolic_fluxes!(flux_parabolic,
                                              i, j, k, element)
             gradients_3_node = get_node_vars(gradients_z, equations_parabolic, dg,
                                              i, j, k, element)
+            x_node = get_node_coords(node_coordinates, equations_parabolic, dg,
+                                     i, j, k, element)
 
             # Calculate parabolic flux and store each component for later use
             flux_parabolic_node_x = flux(u_node,
                                          (gradients_1_node, gradients_2_node,
-                                          gradients_3_node), 1, equations_parabolic)
+                                          gradients_3_node), 1, x_node, t,
+                                         equations_parabolic)
             flux_parabolic_node_y = flux(u_node,
                                          (gradients_1_node, gradients_2_node,
-                                          gradients_3_node), 2, equations_parabolic)
+                                          gradients_3_node), 2, x_node, t,
+                                         equations_parabolic)
             flux_parabolic_node_z = flux(u_node,
                                          (gradients_1_node, gradients_2_node,
-                                          gradients_3_node), 3, equations_parabolic)
+                                          gradients_3_node), 3, x_node, t,
+                                         equations_parabolic)
             set_node_vars!(flux_parabolic_x, flux_parabolic_node_x,
                            equations_parabolic, dg,
                            i, j, k, element)